Fix for VMware Fusion Driver
When configuring a Docker machine for NFS mounts, the IP address in this
script differed from the IP given to my VMware VM. I've updated the
script to use the Docker machine IP, instead of 8.8.8.8. This causes the
`route get` command to return the proper `vmnet<#>` device instead of
whatever is tied to 8.8.8.8.

Also fixed the README.md file with the new `-i` option.